Default arguments added to get methods
All of the get methods have had the Optional default argument added to
allow the return of a default value should the method not return a
value. This will suppress the keyNotFoundException but does not
indicate whether or not the key was found when executing the method.
The preferred method is to use HasKey prior to calling the get method.